requirements.txt
# Google APIs
google-cloud-vision>=3.5.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
gspread>=5.12.0

# CLI and scheduling
typer[all]>=0.9.0
apscheduler>=3.10.4

# Utilities
python-dateutil>=2.8.2
rich>=13.7.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Image processing fallback
Pillow>=10.1.0
